+++ memote/suite/tests/test_basic.py
@@ -91,33 +91,6 @@ def test_metabolites_presence(read_only_model):
     assert len(ann["data"]) >= 1, ann["message"]
 
 
-@annotate(title="Total Number of Transport Reactions", type="count")
-def test_transport_reaction_presence(read_only_model):
-    """
-    Expect more than one transport reaction to be defined in the model.
-
-    Cellular metabolism in any organism usually involves the transport of
-    metabolites across a lipid bi-layer. Hence, this test checks that there is
-    at least one reaction, which transports metabolites from one compartment
-    to another.
-
-    A transport reaction is defined as follows:
-    1. It contains metabolites from at least two compartments and
-    2. at least one metabolite undergoes no chemical conversion, i.e.,
-    the formula stays the same on both sides of the equation.
-
-    This test will not be able to identify transport via the PTS System.
-    """
-    ann = test_transport_reaction_presence.annotation
-    ann["data"] = get_ids(helpers.find_transport_reactions(read_only_model))
-    ann["message"] = wrapper.fill(
-        """A total of {:d} ({:.2%}) transport reactions are defined in the
-        model, this excludes purely metabolic reactions, exchanges, or
-        pseudo-reactions: {}""".format(
-            len(ann["data"]), ann["metric"], truncate(ann["data"])))
-    assert len(ann["data"]) >= 1, ann["message"]
-
-
 @annotate(title="Metabolites without Formula", type="count")
 def test_metabolites_formula_presence(read_only_model):
     """
@@ -315,7 +288,38 @@ def test_find_constrained_pure_metabolic_reactions(read_only_model):
 
 @annotate(title="Number of Transport Reactions", type="count")
 def test_find_transport_reactions(read_only_model):
-    """Expect >= 1 transport reactions are present in the read_only_model."""
+    """
+    Expect >= 1 transport reactions are present in the read_only_model.
+
+    Cellular metabolism in any organism usually involves the transport of
+    metabolites across a lipid bi-layer. This test reports how many
+    of these reactions, which transports metabolites from one compartment
+    to another, are present in the model, as at least one transport reaction
+    must be present for cells to take up nutrients and/or excrete waste.
+
+    A transport reaction is defined as follows:
+    1. It contains metabolites from at least 2 compartments and
+    2. at least 1 metabolite undergoes no chemical reaction, i.e.,
+    the formula and/or annotation stays the same on both sides of the equation.
+
+    A notable exception is transport via PTS, which also contains the following
+    restriction:
+    3. The transported metabolite(s) are transported into a compartment through
+    the exchange of a phosphate.
+
+    An example of tranport via PTS would be
+    pep(c) + glucose(e) -> glucose-6-phosphate(c) + pyr(c)
+
+    Reactions similar to transport via PTS (referred to as "modified transport
+    reactions") follow a similar pattern:
+    A(x) + B-R(y) -> A-R(y) + B(y)
+
+    Such modified transport reactions can be detected, but only when a formula
+    field exists for all metabolites in a particular reaction. If this is not
+    the case, transport reactions are identified through annotations, which
+    cannot detect modified tranport reactions.
+
+    """
     ann = test_find_transport_reactions.annotation
     ann["data"] = get_ids(helpers.find_transport_reactions(read_only_model))
     ann["metric"] = len(ann["data"]) / len(read_only_model.reactions)
@@ -339,11 +343,26 @@ def test_find_constrained_transport_reactions(read_only_model):
     'pass' criteria.
 
     A transport reaction is defined as follows:
-    1. It contains metabolites from at least two compartments and
-    2. at least one metabolite undergoes no chemical conversion, i.e.,
-    the formula stays the same on both sides of the equation.
+    1. It contains metabolites from at least 2 compartments and
+    2. at least 1 metabolite undergoes no chemical reaction, i.e.,
+    the formula and/or annotation stays the same on both sides of the equation.
 
-    This test will not be able to identify transport via the PTS System.
+    A notable exception is transport via PTS, which also contains the following
+    restriction:
+    3. The transported metabolite(s) are transported into a compartment through
+    the exchange of a phosphate.
+
+    An example of tranport via PTS would be
+    pep(c) + glucose(e) -> glucose-6-phosphate(c) + pyr(c)
+
+    Reactions similar to transport via PTS (referred to as "modified transport
+    reactions") follow a similar pattern:
+    A(x) + B-R(y) -> A-R(y) + B(y)
+
+    Such modified transport reactions can be detected, but only when a formula
+    field exists for all metabolites in a particular reaction. If this is not
+    the case, transport reactions are identified through annotations, which
+    cannot detect modified tranport reactions.
     """
     ann = test_find_constrained_transport_reactions.annotation
     transporters = helpers.find_transport_reactions(read_only_model)